pyproject.toml
[tool.poetry]
name = "test_token"
version = "0.1.0"
description = ""
authors = ["Stanislav Glazko"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.7"
djangorestframework = "^3.12.2"
djoser = "^2.1.0"
djangorestframework-simplejwt = "^4.6.0"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
gunicorn = "^20.0.4"
flake8 = "^3.8.4"
requests = "^2.25.1"
pytest-cov = "^2.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
